Adds debug functionality to every class
every class(except Driver) now has a method <private void debug(String
s)>, this will write the class and the input to a log file. Every class
also has new constructors (one for each constructor the class already
had) that takes a BufferedWriter in, this constructor must be used for
debug to work.